Removed CategoryManager
