+++ src/Pim/Bundle/CatalogBundle/Resources/config/managers.yml
@@ -2,7 +2,6 @@ parameters:
     pim_catalog.manager.attribute.class:           Pim\Bundle\CatalogBundle\Manager\AttributeManager
     pim_catalog.manager.attribute_option.class:    Pim\Bundle\CatalogBundle\Manager\AttributeOptionManager
     pim_catalog.manager.attribute_group.class:     Pim\Bundle\CatalogBundle\Manager\AttributeGroupManager
-    pim_catalog.manager.category.class:            Pim\Bundle\CatalogBundle\Manager\CategoryManager
     pim_catalog.manager.channel.class:             Pim\Bundle\CatalogBundle\Manager\ChannelManager
     pim_catalog.manager.completeness.class:        Pim\Bundle\CatalogBundle\Manager\CompletenessManager
     pim_catalog.manager.currency.class:            Pim\Bundle\CatalogBundle\Manager\CurrencyManager
@@ -29,14 +28,6 @@ services:
         arguments:
             - '@pim_catalog.repository.product_category'
             - '@pim_catalog.repository.category'
-
-    pim_catalog.manager.category:
-        class: %pim_catalog.manager.category.class%
-        arguments:
-            - '@doctrine.orm.entity_manager'
-            - '@pim_catalog.repository.category'
-            - '@pim_catalog.factory.category'
-            - %pim_catalog.entity.category.class%
 
     pim_catalog.manager.attribute:
         class: %pim_catalog.manager.attribute.class%
