+++ src/Pim/Bundle/EnrichBundle/Resources/config/datagrid/channel.yml
@@ -67,4 +67,4 @@ datagrid:
                     options:
                         field_options:
                             multiple: true
-                            choices:  '@pim_catalog.manager.category->getTreeChoices'
+                            choices:  '@pim_catalog.repository.category->getTreeChoices'
